move ApplicationPage
